Build crosvm+deps for aarch64 on x86-hosted docker
To build aarch64 binaries using docker on an x86 host:

./device/google/cuttlefish_vmm/rebuild.sh --docker --docker_arch aarch64

This will build a container image for aarch64; start it; and then log
into it with the same $USER as that of the caller, in order to build
crosvm and its dependencies.  By default, the output will be saved in
./device/google/cuttlefish_vmm/aarch64-linux-gnu/.

If you develop for aarch64 and want to use this script to build crosvm
as you develop, map the sources when you start the container, using
--docker_source.  Since the docker-image build takes a while for
aarch64, assuming you've built the docker image already, then you can
pass the option --nodocker_build_image to simply create a new container:

./device/google/cuttlefish_vmm/rebuild.sh \
	--docker --docker_arch aarch64 \
	--nodocker_build_image \
	--docker_source /path/to/source

This will skip the image-build step; it will create a new container from
scratch, sync the sources under /path/to/source (even if they were
synced there previously), build them, and copy the output.

On subsequent rebuilds, you can pass the --reuse flag to skip the image
biuld, container creation, and syncing of the sources, simply to rebuld
them.  This is the fastest way to use the script iteratively:

vim /path/to/source/some/file.rs
'#' hack hack hack
./device/google/cuttlefish_vmm/rebuild.sh \
	--docker --docker_arch aarch64 \
	--reuse

Note that you will need to move the built objects manually to your ARM
board before you can use them.

Bug: 148642775 Clean up the crosvm build
Test: built using gce, arm, docker targeting x86 and arm64

Change-Id: I36c31e77787b0990b9b537ad54b4dcfc3d5fb829
Signed-off-by: Iliyan Malchev <malchev@google.com>