Introduce --[no]docker_persistent
Option --docker_persistent, enabled by default, allows you to build
crosvm from a non-persistent, non-privileged container.  This might be
better for automated builds on a build server that does not want to
allow privileged containers.

	./device/google/cuttlefish_vmm/rebuild.sh \
		--docker --nodocker_persistent

The down side is that the container is removed after each run, and
option --reuse only skips the image-building step.

This patch also modifies the --gce and --arm usecases to invoke the
*_rebuild (aarch64_rebuild and x86_64_rebuild) targets of
rebuild-internal.sh when --reuse is specified.  Previously, the
aarc64_build and x86_64_build targets were used even with --reuse, which
caused the working directory to be wiped out prior to starting the build
each time.

Finally, we fix a bug with the usage of --docker_source and
--docker_working command-line options.

Bug: 148642775 Clean up the crosvm build
Test: run rebuild.sh for --gce, --arm, --docker usecases

Change-Id: I66a9dc8971b20f5288aea2972937e031936db312
Signed-off-by: Iliyan Malchev <malchev@google.com>