Build crosvm only for aarch64
We have moved the crosvm build for x86_64 to be built natively by the
Android build system, and have deleted the prebuilts, so we should no
longer encourage use of the docker rebuild script to build an x86_64
version.

We still need the script to build the gfxstream binaries for x86_64
hermetically, and to build everything for aarch64, so this just makes
the necessary tweak to disable crosvm for x86_64.

Bug: 161831173
Test: ./rebuild.sh --docker
Test: ./rebuild.sh --docker --docker_arch aarch64
Change-Id: Iad747ab007183dac33f7c1edf44c843371948d02