Remove separated aarch64/x86_64 manifests
These have been required to be updated in tandem for a while. Moving
forward, we will be getting crosvm for x86_64 from the Android tree,
so it will be a bug to have a different build configuration for
that build and the docker builds. Remove the manifest duplication.

NOTE: This change removes the gfxstream repos, but they are brought
      back in the next change.

Bug: 161831173
Change-Id: I6055f8594caee1d0e5131aa69b8273e021534bb2