common_crosvm is build as sh_binary_host
Previously, it was built using cc_prebuilt_binary. The problem with that
was that what's installed to out/soong/host is actually a symlink to the
source (e.g.  /home/<username>/<androidroot>/foo/bar/...). See
I1a600c7163ce0ec34ee8caf0ffe87fef4feb3064 for more info about the
symlink.

Since cc_prebuilt_binary is actually for 'compiled' binaries, switch to
sh_binary_host which is exactly for shell scripts. These modules are
installed not as symlinks.

This change also adds the implementation of the script for the
linux_bionic_* targets where the bionic linker has to be used instead of
the host linker (ld-linux-<arch>.so).

Bug: 168086242
Test: m
Change-Id: Ib7743bba0aa43465514b178d95c2cdf7a0950de4