Update crosvm aarch64 prebuilt & build scripts
Adds a workaround to the build scripts to accomodate for aosp/1412815.
That change builds the rust protobuf sources outside of the cargo build
system, so this change updates the scripts to do something comparable
before starting the cargo build.

Unfortunately, the workaround is a noticeable slowdown in the process as
it builds and installs the protobuf source to rust source compiler.

Test: Treehugger
Bug: 155695742
Change-Id: I4abcd86084b19eb3f19e37955fd017e98cab2eb1