Remove prebuilt modules for crosvm/arm64
Soong prebuilt module definitions for the crosvm/arm64 binaries are
removed because they are now built from source code.

common_crosvm script is moved to external/crosvm because that's a better
place than here.

gen_android_bp.sh was slightly modified to not create the deleted Soong
module definitions.

Finally, the existing prebuilts are removed because nobody uses it
anymore.

Bug: 159685774
Test: m cvd-host_package
Change-Id: I24a82fac803d4d9d1a6496738254324e1cc76495