Modify minijail-sys to link dynamically
When minijail-sys is built with cargo, it would link libminijail.pic.a
which ultimately causes this library to be linked into crosvm
statically.

When minijail-sys is built by the Android build system, it is linked
dynamically to libminijail.so, which makes it easier to share the
library and to debug issues in it.

Change the cargo build process to use a dynamically linked object.

Bug: 246436720
Change-Id: Ida49760eeb9e02dbde1edecedf397cbdd7e1ac29